Data.LCA.Online.Polymorphic

Description

Provides online calculation of the the lowest common ancestor in O(log h) by compressing the spine of the paths using a skew-binary random access list.

This library implements the technique described in my talk

http://www.slideshare.net/ekmett/skewbinary-online-lowest-common-ancestor-search

to improve the known asymptotic bounds on both online lowest common ancestor search

http://en.wikipedia.org/wiki/Lowest_common_ancestor

and the online level ancestor problem:

http://en.wikipedia.org/wiki/Level_ancestor_problem

Algorithms used here assume that the key values chosen for k are globally unique.

This version provides access to a monoidal "summary" of the elided path for many operations.

Documentation

data Path k a

A compressed Path as a skew binary random access list

toList :: Path k a -> [(k, a)]

Convert a Path to a list of (ID, value) pairs.

fromList :: Monoid a => [(k, a)] -> Path k a

Build a Path from a list of (ID, value) pairs.

map :: Monoid b => (a -> b) -> Path k a -> Path k b

O(n) Re-annotate a Path full of monoidal values using a different Monoid.

mapHom :: (a -> b) -> Path k a -> Path k b

O(n) Re-annotate a Path full of monoidal values/

Unlike map, mapHom f assumes that f is a Monoid homomorphism, that is to say you must ensure

f a `mappend' f b = f (a `mappend' b)
f mempty = mempty

mapWithKey :: Monoid b => (k -> a -> b) -> Path k a -> Path k b

O(n) Re-annotate a Path full of monoidal values with access to the key.

traverseP :: (Applicative f, Monoid b) => (a -> f b) -> Path k a -> f (Path k b)

Traverse a Path yielding a new monoidal annotation.

traverseWithKey :: (Applicative f, Monoid b) => (k -> a -> f b) -> Path k a -> f (Path k b)

Traverse a Path with access to the node IDs.

empty :: Path k a

The empty Path

cons :: Monoid a => k -> a -> Path k a -> Path k a

O(1) Invariant: most operations assume that the keys k are globally unique

Extend the Path with a new node ID and value.

uncons :: Monoid a => Path k a -> Maybe (k, a, Path k a)

O(1) Extract the node ID and value from the newest node on the Path.

view :: Monoid a => Path k a -> View k Path a

O(1) Extract the node ID and value from the newest node on the Path, slightly faster than uncons.

null :: Foldable t => t a -> Bool

Test whether the structure is empty. The default implementation is optimized for structures that are similar to cons-lists, because there is no general way to do better.

length :: Foldable t => t a -> Int

Returns the size/length of a finite structure as an Int. The default implementation is optimized for structures that are similar to cons-lists, because there is no general way to do better.

measure :: Monoid a => Path k a -> a

Extract a monoidal summary of a Path.

isAncestorOf :: (Monoid b, Eq k) => Path k a -> Path k b -> Bool

O(log h) xs `isAncestorOf' ys holds when xs is a prefix starting at the root of path ys.

keep :: Monoid a => Int -> Path k a -> Path k a

O(log (h - k)) to keep k elements of Path of length h

This solves the online version of the "level ancestor problem" with no preprocessing in O(log h) time, improving known complexity bounds.

http://en.wikipedia.org/wiki/Level_ancestor_problem

mkeep :: (Monoid a, Monoid b) => (a -> b) -> Int -> Path k a -> (b, Path k a)

O(log (h - k)) to keep k elements of Path of length h, and provide a monoidal summary of the dropped elements using a supplied monoid homomorphism.

drop :: Monoid a => Int -> Path k a -> Path k a

O(log k) to drop k elements from a Path

mdrop :: (Monoid a, Monoid b) => (a -> b) -> Int -> Path k a -> (b, Path k a)

O(log k) to drop k elements from a Path and provide a monoidal summary of the dropped elements using a suplied monoid homomorphism

(~=) :: Eq k => Path k a -> Path k b -> Bool infix 4

O(1) Compare to see if two trees have the same leaf key

lca :: (Monoid a, Monoid b, Eq k) => Path k a -> Path k b -> Path k a

O(log h) Compute the lowest common ancestor of two paths

mlca :: (Monoid a, Monoid b, Monoid c, Monoid d, Eq k) => (a -> c) -> (b -> d) -> Path k a -> Path k b -> (c, Path k a, d, Path k b)

O(log h) Compute the lowest common ancestor of two paths along with a monoidal summary of their respective tails using the supplied monoid homomorphisms.

data View k f a

Provides a consistent View for peeling off the bottom node of a path.

Constructors

Root

Node !k a (f k a)

nearest :: (Foldable f, Monoid a, Monoid b, Eq k) => Path k b -> f (Path k a) -> Maybe (Path k a)

concatPaths :: Monoid a => Path k a -> Path k a -> Path k a