Tezos.Common.Json

Documentation

parseAsString :: forall a. (Read a, Typeable a) => Value -> Parser a

parseStringEncodedIntegral :: (Read a, Integral a, Typeable a) => Value -> Parser (StringEncode a)

deriveTezosJson :: Name -> Q [Dec]

deriveTezosToJson :: Name -> Q [Dec]

deriveTezosJsonKind :: String -> Name -> Q [Dec]

tezosJsonOptions :: Options

tezosJsonOptionsKind :: String -> Options

newtype StringEncode a

JavaScript compatibility forces us and the Tezos JSON/RPC to encode any integer with more than 32 bits into a decimal string format. StringEncode helps us safely represent a range of different types for which we must always respect this property.

Constructors

StringEncode a

type TezosInt64 = StringEncode Int64

type TezosBigNum = StringEncode Integer

type TezosPositiveBigNum = StringEncode Natural

WARNING: Due to a major oversight on the part of Tezos Core, this type is incorrectly named. It represents _non-negative_ integers (including zero), not _strictly positive_ integers. Therefore the Haskell Natural type is indeed the correct counterpart.

newtype JsonRpcError

"$ref": "#definitionserror",

Constructors

JsonRpcError Value