Tezos.V004.Block

Documentation

data Block

"description": "All the information about a block.",

Constructors

Block

Fields

_block_protocol :: !ProtocolHash

_block_chainId :: !ChainId

_block_hash :: !BlockHash

_block_header :: !BlockHeaderFull

_block_metadata :: !BlockMetadata

_block_operations :: !(Seq (Seq Operation))

data MaxOperationListLength

Constructors

MaxOperationListLength

Fields

_maxOperationListLength_maxSize :: !Int

_maxOperationListLength_maxOp :: !(Maybe Int)

data VotingPeriodKind

"voting_period_kind": {

Constructors

VotingPeriodKind_Proposal

VotingPeriodKind_TestingVote

VotingPeriodKind_Testing

VotingPeriodKind_PromotionVote

data BlockMetadata

"block_header_metadata": {

Constructors

BlockMetadata

Fields

_blockMetadata_protocol :: !ProtocolHash

_blockMetadata_nextProtocol :: !ProtocolHash

_blockMetadata_testChainStatus :: !TestChainStatus

_blockMetadata_maxOperationsTtl :: !Int

_blockMetadata_maxOperationDataLength :: !Int

_blockMetadata_maxBlockHeaderLength :: !Int

_blockMetadata_maxOperationListLength :: !(Seq MaxOperationListLength)

_blockMetadata_baker :: !PublicKeyHash

_blockMetadata_level :: !Level

_blockMetadata_votingPeriodKind :: !VotingPeriodKind

_blockMetadata_nonceHash :: !(Maybe CycleNonce)

_blockMetadata_consumedGas :: !TezosPositiveBigNum

_blockMetadata_deactivated :: !(Seq PublicKeyHash)

_blockMetadata_balanceUpdates :: !(Seq BalanceUpdate)

data MonitorBlock

Constructors

MonitorBlock

Fields

_monitorBlock_hash :: !BlockHash

_monitorBlock_level :: !RawLevel

_monitorBlock_proto :: !Word8

_monitorBlock_predecessor :: !BlockHash

_monitorBlock_timestamp :: !UTCTime

_monitorBlock_validationPass :: !Word8

_monitorBlock_operationsHash :: !OperationListListHash

_monitorBlock_fitness :: !Fitness

_monitorBlock_context :: !ContextHash

data VeryBlockLike

Constructors

VeryBlockLike

Fields

_veryBlockLike_hash :: !BlockHash

_veryBlockLike_predecessor :: !BlockHash

_veryBlockLike_fitness :: !Fitness

_veryBlockLike_level :: !RawLevel

_veryBlockLike_timestamp :: !UTCTime

toBlockHeader :: Block -> BlockHeader

data WithProtocolHash a

Simple wrapper for adding a protocol hash to some other value.

The Aeson instances are smart (too smart). If the underlying type is not an object or an object with a conflicting key with the one added by this type, the JSON encoding will be two-layered. Otherwise, the JSON encoding will simply add an additional key for the protocol information.

Constructors

WithProtocolHash

Fields

_withProtocolHash_value :: !a

_withProtocolHash_protocolHash :: !ProtocolHash

withProtocolHash_value :: forall a a. Lens (WithProtocolHash a) (WithProtocolHash a) a a

withProtocolHash_protocolHash :: forall a. Lens' (WithProtocolHash a) ProtocolHash

veryBlockLike_timestamp :: Lens' VeryBlockLike UTCTime

veryBlockLike_predecessor :: Lens' VeryBlockLike BlockHash

veryBlockLike_level :: Lens' VeryBlockLike RawLevel

veryBlockLike_hash :: Lens' VeryBlockLike BlockHash

veryBlockLike_fitness :: Lens' VeryBlockLike Fitness

monitorBlock_validationPass :: Lens' MonitorBlock Word8

monitorBlock_timestamp :: Lens' MonitorBlock UTCTime

monitorBlock_proto :: Lens' MonitorBlock Word8

monitorBlock_predecessor :: Lens' MonitorBlock BlockHash

monitorBlock_operationsHash :: Lens' MonitorBlock OperationListListHash

monitorBlock_level :: Lens' MonitorBlock RawLevel

monitorBlock_hash :: Lens' MonitorBlock BlockHash

monitorBlock_fitness :: Lens' MonitorBlock Fitness

monitorBlock_context :: Lens' MonitorBlock ContextHash

maxOperationListLength_maxSize :: Lens' MaxOperationListLength Int

maxOperationListLength_maxOp :: Lens' MaxOperationListLength (Maybe Int)

blockMetadata_votingPeriodKind :: Lens' BlockMetadata VotingPeriodKind

blockMetadata_testChainStatus :: Lens' BlockMetadata TestChainStatus

blockMetadata_protocol :: Lens' BlockMetadata ProtocolHash

blockMetadata_nonceHash :: Lens' BlockMetadata (Maybe CycleNonce)

blockMetadata_nextProtocol :: Lens' BlockMetadata ProtocolHash

blockMetadata_maxOperationsTtl :: Lens' BlockMetadata Int

blockMetadata_maxOperationListLength :: Lens' BlockMetadata (Seq MaxOperationListLength)

blockMetadata_maxOperationDataLength :: Lens' BlockMetadata Int

blockMetadata_maxBlockHeaderLength :: Lens' BlockMetadata Int

blockMetadata_level :: Lens' BlockMetadata Level

blockMetadata_deactivated :: Lens' BlockMetadata (Seq PublicKeyHash)

blockMetadata_consumedGas :: Lens' BlockMetadata TezosPositiveBigNum

blockMetadata_balanceUpdates :: Lens' BlockMetadata (Seq BalanceUpdate)

blockMetadata_baker :: Lens' BlockMetadata PublicKeyHash

block_protocol :: Lens' Block ProtocolHash

block_operations :: Lens' Block (Seq (Seq Operation))

block_metadata :: Lens' Block BlockMetadata

block_header :: Lens' Block BlockHeaderFull

block_hash :: Lens' Block BlockHash

block_chainId :: Lens' Block ChainId

class BlockLike b where

Methods

hash :: Lens' b BlockHash

predecessor :: Lens' b BlockHash

level :: Lens' b RawLevel

fitness :: Lens' b Fitness

timestamp :: Lens' b UTCTime

class HasProtocolHash a where

Methods

protocolHash :: Lens' a ProtocolHash

class HasChainId a where

Methods

chainIdL :: Lens' a ChainId

class HasBlockMetadata a where

Methods

blockMetadata :: Lens' a BlockMetadata

class HasBlockHeaderFull a where

Methods

blockHeaderFull :: Lens' a BlockHeaderFull

mkVeryBlockLike :: BlockLike b => b -> VeryBlockLike