Tezos.V004.Fitness

Documentation

newtype FitnessF a

Constructors

FitnessF

Fields

unFitnessF :: Seq a

type Fitness' a = FitnessF (Base16ByteString a)

for these to be useful, you'd need `TezosBinary ByteString`, but that's definately not the same one as needed for the above FromJSON instances; since both would be needed, both are wrong. we satisfy ourselves with the fully monomorphic instances for both

instance TezosBinary a => TezosBinary (FitnessF a) where parseBinary = FitnessF $ parseBinary encodeBinary = encodeBinary . unFitnessF

type Fitness = Fitness' ShortByteString

toFitness :: Seq a -> Fitness' a

unFitness :: FitnessF (Base16ByteString a) -> Seq a