Tezos.V004.Unsafe

Documentation

unsafeAssumptionLevelToCycle :: ProtoInfo -> RawLevel -> Cycle

Convert a level to the cycle that contains it.

We subtract 1 because the first cycle begins after the genesis block. Yet while that block is not part of the cycle, it is still given a level, level 0.

unsafeAssumptionFirstLevelInCycle :: ProtoInfo -> Cycle -> RawLevel

Convert a cycle to the level of the first block in that cycle.

We add 1 because we do not consider the genesis block as part of the first cycle, as that would make the first cycle alone 1 block larger than all the others.

unsafeAssumptionRightsContextLevel :: ProtoInfo -> RawLevel -> RawLevel

We want the first block in the cycle that sits PRESERVED_CYCLES before the requested level, that is on the correct branch.