Tezos.V005.Account

Documentation

type ManagerKey = Maybe PublicKey

data Account

Constructors

Account

Fields

_account_delegate :: !(Maybe PublicKeyHash)

_account_balance :: !Tez

_account_script :: !(Maybe ContractScript)

_account_counter :: !(Maybe Counter)

account_balance :: Lens' Account Tez

account_counter :: Lens' Account (Maybe Counter)

account_delegate :: Lens' Account (Maybe PublicKeyHash)

account_script :: Lens' Account (Maybe ContractScript)

data DelegateInfo

Constructors

DelegateInfo

Fields

_delegateInfo_balance :: !Tez

_delegateInfo_frozenBalance :: !Tez

_delegateInfo_frozenBalanceByCycle :: !(Seq FrozenBalanceByCycle)

_delegateInfo_stakingBalance :: !Tez

_delegateInfo_delegatedContracts :: !(Seq ContractId)

_delegateInfo_delegatedBalance :: !Tez

_delegateInfo_deactivated :: !Bool

_delegateInfo_gracePeriod :: !Cycle

data FrozenBalanceByCycle

Constructors

FrozenBalanceByCycle

Fields

_frozenBalanceByCycle_cycle :: !Cycle

_frozenBalanceByCycle_deposit :: !Tez

_frozenBalanceByCycle_fees :: !Tez

_frozenBalanceByCycle_rewards :: !Tez

data EndorsingRights

Constructors

EndorsingRights

Fields

_endorsingRights_level :: !RawLevel

_endorsingRights_delegate :: !PublicKeyHash

_endorsingRights_slots :: !(Seq Word8)

_endorsingRights_estimatedTime :: !(Maybe UTCTime)

data BakingRights

Constructors

BakingRights

Fields

_bakingRights_level :: !RawLevel

_bakingRights_delegate :: !PublicKeyHash

_bakingRights_priority :: !Priority

_bakingRights_estimatedTime :: !(Maybe UTCTime)

data AccountDelegate

Constructors

AccountDelegate

Fields

_accountDelegate_setable :: !Bool

_accountDelegate_value :: !(Maybe PublicKeyHash)

type Counter = TezosPositiveBigNum

A special nonce that must increment with each successive operation signed and submitted by a particular account.

account_manager :: Lens' Account PublicKeyHash

account_spendable :: Lens' Account Bool

accountDelegate_setable :: Lens' AccountDelegate Bool

accountDelegate_value :: Lens' AccountDelegate (Maybe PublicKeyHash)

bakingRights_delegate :: Lens' BakingRights PublicKeyHash

bakingRights_estimatedTime :: Lens' BakingRights (Maybe UTCTime)

bakingRights_level :: Lens' BakingRights RawLevel

bakingRights_priority :: Lens' BakingRights Priority

delegateInfo_balance :: Lens' DelegateInfo Tez

delegateInfo_deactivated :: Lens' DelegateInfo Bool

delegateInfo_delegatedBalance :: Lens' DelegateInfo Tez

delegateInfo_delegatedContracts :: Lens' DelegateInfo (Seq ContractId)

delegateInfo_frozenBalance :: Lens' DelegateInfo Tez

delegateInfo_frozenBalanceByCycle :: Lens' DelegateInfo (Seq FrozenBalanceByCycle)

delegateInfo_gracePeriod :: Lens' DelegateInfo Cycle

delegateInfo_stakingBalance :: Lens' DelegateInfo Tez

endorsingRights_delegate :: Lens' EndorsingRights PublicKeyHash

endorsingRights_estimatedTime :: Lens' EndorsingRights (Maybe UTCTime)

endorsingRights_level :: Lens' EndorsingRights RawLevel

endorsingRights_slots :: Lens' EndorsingRights (Seq Word8)

frozenBalanceByCycle_cycle :: Lens' FrozenBalanceByCycle Cycle

frozenBalanceByCycle_deposit :: Lens' FrozenBalanceByCycle Tez

frozenBalanceByCycle_fees :: Lens' FrozenBalanceByCycle Tez

frozenBalanceByCycle_rewards :: Lens' FrozenBalanceByCycle Tez