Tezos.V005.Micheline

Documentation

data MichelinePrimAp

Constructors

MichelinePrimAp

Fields

_michelinePrimAp_prim :: !MichelinePrimitive

_michelinePrimAp_args :: !(Seq Expression)

_michelinePrimAps_annots :: !(Seq Annotation)

data Annotation

Constructors

Annotation_Type !Text

Annotation_Variable !Text

Annotation_Field !Text

data Expression

Constructors

Expression_Int !TezosBigNum
Micheline represents both nats and ints using the same decimal format. The Haskell Integer type spans all possible values that the final (Michelson) type could end up being, and then some, so we use (StringEncode Integer) to represent all integral values here for easy JSON encoding compatibility.

Expression_String !Text

Expression_Bytes !(Base16ByteString ByteString)

Expression_Seq !(Seq Expression)

Expression_Prim !MichelinePrimAp

newtype MichelinePrimitive

Constructors

MichelinePrimitive Text

michelineV1Primitive :: Seq Text

michelsonV1EnumV005 :: Seq Text

michelsonV005ChangesEnum :: Seq Text

annotFromText :: MonadFail m => Text -> m Annotation

annotToText :: Annotation -> Text

putAnnotationSeq :: Seq Annotation -> Const Builder ()

getAnnotationSeq :: Get (Seq Annotation)

michelsonV1Enum :: Seq Text