Tezos.V005.Operation

Documentation

data Operation

"operation": {

Constructors

Operation

Fields

_operation_protocol :: !ProtocolHash

_operation_chainId :: !ChainId

_operation_hash :: !OperationHash

_operation_branch :: !BlockHash

_operation_contents :: !(Seq OperationContents)

_operation_signature :: !(Maybe Signature)

data OperationWithMetadata

Constructors

OperationWithMetadata

Fields

_operationWithMetadata_contents :: !(Seq OperationContents)

_operationWithMetadata_signature :: !(Maybe Signature)

data OpKind

Constructors

OpKind_SeedNonceRevelation

OpKind_DoubleEndorsementEvidence

OpKind_DoubleBakingEvidence

OpKind_ActivateAccount

OpKind_Endorsement

OpKind_Proposals

OpKind_Ballot

OpKind_Manager [OpKindManager]

data OpKindTag opKind where

Constructors

OpKindTag_Endorsement :: OpKindTag OpKind_Endorsement

OpKindTag_SeedNonceRevelation :: OpKindTag OpKind_SeedNonceRevelation

OpKindTag_DoubleEndorsementEvidence :: OpKindTag OpKind_DoubleEndorsementEvidence

OpKindTag_DoubleBakingEvidence :: OpKindTag OpKind_DoubleBakingEvidence

OpKindTag_ActivateAccount :: OpKindTag OpKind_ActivateAccount

OpKindTag_Proposals :: OpKindTag OpKind_Proposals

OpKindTag_Ballot :: OpKindTag OpKind_Ballot

OpKindTag_Manager :: OpKindManagerTag opKindManager -> OpKindTag (OpKind_Manager (opKindManager ': '[]))

data OpsKindTag opKinds where

Constructors

OpsKindTag_Single :: OpKindTag k -> OpsKindTag k

OpsKindTag_Cons :: OpKindTag (OpKind_Manager (k ': '[])) -> OpsKindTag (OpKind_Manager (kk ': ks)) -> OpsKindTag (OpKind_Manager (k ': (kk ': ks)))

data OpKindManager

Constructors

OpKindManager_Reveal

OpKindManager_Transaction

OpKindManager_Origination

OpKindManager_Delegation

data OpKindManagerTag opKindManager where

Constructors

OpKindManagerTag_Reveal :: OpKindManagerTag OpKindManager_Reveal

OpKindManagerTag_Transaction :: OpKindManagerTag OpKindManager_Transaction

OpKindManagerTag_Origination :: OpKindManagerTag OpKindManager_Origination

OpKindManagerTag_Delegation :: OpKindManagerTag OpKindManager_Delegation

data BatchableOp k where

Constructors

BatchableOp :: OpKindManagerTag k -> BatchableOp (OpKind_Manager '[k])

data Op (a :: OpKind)

Constructors

Op

Fields

_op_branch :: !BlockHash

_op_contents :: !(OpContentsList a)

_op_signature :: !(Maybe Signature)

data OpWithChain a

Constructors

OpWithChain

Fields

_opWithChain_operation :: !a

_opWithChain_chain_id :: !ChainId

data ProtoOp (a :: OpKind)

next_operation: { "protocol": Pt24m4xiPbLDhVgVfABUjirbmda3yohdN82Sp9FeuAXJ4eV9otd, "branch": $block_hash, "contents": [ $operation.alpha.contents ... ], "signature": $Signature }

Like Op but with a protocol hash added. Used in preapply.

Constructors

ProtoOp

Fields

_protoOp_protocol :: !ProtocolHash

_protoOp_branch :: !BlockHash

_protoOp_contents :: !(OpContentsList a)

_protoOp_signature :: !(Maybe Signature)

stripOpProtocol :: ProtoOp a -> Op a

data PendingOp (a :: OpKind)

Like Op but with its own hash. Used in the mempool.

Constructors

PendingOp

Fields

_pendingOp_hash :: !OperationHash

_pendingOp_branch :: !BlockHash

_pendingOp_contents :: !(OpContentsList a)

_pendingOp_signature :: !Signature

data ErroredOp (a :: OpKind)

This is also used in the mempool for operations that will not be applied for one reason or another.

Constructors

ErroredOp

Fields

_erroredOp_protocol :: !ProtocolHash

_erroredOp_branch :: !BlockHash

_erroredOp_contents :: !(OpContentsList a)

_erroredOp_signature :: !Signature

_erroredOp_error :: !Value

data EmptyMetadata

Constructors

EmptyMetadata

data NoContextOperation

operation.alpha.operation_with_metadata { "contents": [ $operation.alpha.operation_contents_and_result ... ], "signature"?: $Signature } || { "contents": [ $operation.alpha.contents ... ], "signature"?: $Signature }

This is the return value of the run_operation RPC.

Constructors

NoContextOperation

Fields

_noContextOperation_contents :: !(Seq OperationContents)

_noContextOperation_signature :: !(Maybe Signature)

data OperationContents

"operation.alpha.operation_contents_and_result": {

Constructors

OperationContents_Endorsement !OperationContentsEndorsement

OperationContents_SeedNonceRevelation !OperationContentsSeedNonceRevelation

OperationContents_DoubleEndorsementEvidence !OperationContentsDoubleEndorsementEvidence

OperationContents_DoubleBakingEvidence !OperationContentsDoubleBakingEvidence

OperationContents_ActivateAccount !OperationContentsActivateAccount

OperationContents_Proposals !OperationContentsProposals

OperationContents_Ballot !OperationContentsBallot

OperationContents_Reveal !OperationContentsReveal

OperationContents_Transaction !OperationContentsTransaction

OperationContents_Origination !OperationContentsOrigination

OperationContents_Delegation !OperationContentsDelegation

data OpContentsList (a :: OpKind) where

Constructors

OpContentsList_Single :: OpContents a -> OpContentsList a

OpContentsList_Cons :: OpContents (OpKind_Manager '[k]) -> OpContentsList (OpKind_Manager (kk ': ks)) -> OpContentsList (OpKind_Manager (k ': (kk ': ks)))

data OpContents (a :: OpKind) where

Constructors

OpContents_Endorsement :: !OpContentsEndorsement -> OpContents OpKind_Endorsement

OpContents_SeedNonceRevelation :: !OpContentsSeedNonceRevelation -> OpContents OpKind_SeedNonceRevelation

OpContents_DoubleEndorsementEvidence :: !OpContentsDoubleEndorsementEvidence -> OpContents OpKind_DoubleEndorsementEvidence

OpContents_DoubleBakingEvidence :: !OpContentsDoubleBakingEvidence -> OpContents OpKind_DoubleBakingEvidence

OpContents_ActivateAccount :: !OpContentsActivateAccount -> OpContents OpKind_ActivateAccount

OpContents_Proposals :: !OpContentsProposals -> OpContents OpKind_Proposals

OpContents_Ballot :: !OpContentsBallot -> OpContents OpKind_Ballot

OpContents_Reveal :: !(OpContentsManager OpContentsReveal) -> OpContents (OpKind_Manager '[OpKindManager_Reveal])

OpContents_Transaction :: !(OpContentsManager OpContentsTransaction) -> OpContents (OpKind_Manager '[OpKindManager_Transaction])

OpContents_Origination :: !(OpContentsManager OpContentsOrigination) -> OpContents (OpKind_Manager '[OpKindManager_Origination])

OpContents_Delegation :: !(OpContentsManager OpContentsDelegation) -> OpContents (OpKind_Manager '[OpKindManager_Delegation])

data OpContentsEndorsement

Constructors

OpContentsEndorsement

Fields

_opContentsEndorsement_level :: !RawLevel

data OpContentsSeedNonceRevelation

Constructors

OpContentsSeedNonceRevelation

Fields

_opContentsSeedNonceRevelation_level :: !RawLevel

_opContentsSeedNonceRevelation_nonce :: !(Base16ByteString ByteString)

data OpContentsDoubleEndorsementEvidence

Constructors

OpContentsDoubleEndorsementEvidence

Fields

_opContentsDoubleEndorsementEvidence_op1 :: !InlinedEndorsement

_opContentsDoubleEndorsementEvidence_op2 :: !InlinedEndorsement

data OpContentsDoubleBakingEvidence

Constructors

OpContentsDoubleBakingEvidence

Fields

_opContentsDoubleBakingEvidence_bh1 :: !BlockHeaderFull

_opContentsDoubleBakingEvidence_bh2 :: !BlockHeaderFull

data OpContentsActivateAccount

Constructors

OpContentsActivateAccount

Fields

_opContentsActivateAccount_pkh :: !Ed25519PublicKeyHash

_opContentsActivateAccount_secret :: !(Base16ByteString ByteString)

data OpContentsProposals

Constructors

OpContentsProposals

Fields

_opContentsProposals_source :: !PublicKeyHash

_opContentsProposals_period :: !RawLevel

_opContentsProposals_proposals :: !(Seq ProtocolHash)

data OpContentsBallot

Constructors

OpContentsBallot

Fields

_opContentsBallot_source :: !PublicKeyHash

_opContentsBallot_period :: !RawLevel

_opContentsBallot_proposal :: !ProtocolHash

_opContentsBallot_ballot :: !Ballot

data OpContentsManager op

Constructors

OpContentsManager

Fields

_opContentsManager_source :: !PublicKeyHash

_opContentsManager_fee :: !Tez

_opContentsManager_counter :: !Counter

_opContentsManager_gasLimit :: !TezosPositiveBigNum

_opContentsManager_storageLimit :: !TezosPositiveBigNum

_opContentsManager_operation :: !op

data OpContentsReveal

Constructors

OpContentsReveal

Fields

_opContentsReveal_publicKey :: !PublicKey

newtype EntrypointName

Constructors

EntrypointName

Fields

unEntrypointName :: ByteString

entrypointName :: MonadFail m => ByteString -> m EntrypointName

data Entrypoint

Constructors

EntrypointDefault

EntrypointRoot

EntrypointDo

EntrypointSetDelegate

EntrypointRemoveDelegate

EntrypointOther !EntrypointName

data OpParameters

Constructors

OpParameters

Fields

_opParameters_entrypoint :: !Entrypoint

_opParameters_value :: !Expression

data OpContentsTransaction

Constructors

OpContentsTransaction

Fields

_opContentsTransaction_amount :: !Tez

_opContentsTransaction_destination :: !ContractId

_opContentsTransaction_parameters :: !(Maybe OpParameters)

data OpContentsOrigination

Constructors

OpContentsOrigination

Fields

_opContentsOrigination_balance :: !Tez

_opContentsOrigination_delegate :: !(Maybe PublicKeyHash)

_opContentsOrigination_script :: !ContractScript

data OpContentsDelegation

Constructors

OpContentsDelegation

Fields

_opContentsDelegation_delegate :: !(Maybe PublicKeyHash)

data OperationContentsEndorsement

"kind": { "type": "string", "enum": [ "endorsement" ] },

Constructors

OperationContentsEndorsement

Fields

_operationContentsEndorsement_metadata :: EndorsementMetadata

_operationContentsEndorsement_level :: RawLevel

data EndorsementMetadata

Constructors

EndorsementMetadata

Fields

_endorsementMetadata_balanceUpdates :: !(Seq BalanceUpdate)

_endorsementMetadata_delegate :: !PublicKeyHash

_endorsementMetadata_slots :: !(Seq Word8)

data OperationContentsSeedNonceRevelation

"kind": { "type": "string", "enum": [ "seed_nonce_revelation" ] },

Constructors

OperationContentsSeedNonceRevelation

Fields

_operationContentsSeedNonceRevelation_metadata :: SeedNonceRevelationMetadata

_operationContentsSeedNonceRevelation_level :: RawLevel

_operationContentsSeedNonceRevelation_nonce :: !(Base16ByteString ByteString)

data SeedNonceRevelationMetadata

Constructors

SeedNonceRevelationMetadata

Fields

_seedNonceRevelationMetadata_balanceUpdates :: !(Seq BalanceUpdate)

data InlinedEndorsement

Constructors

InlinedEndorsement

Fields

_inlinedEndorsement_branch :: !BlockHash

_inlinedEndorsement_operations :: !InlinedEndorsementContents

_inlinedEndorsement_signature :: !(Maybe Signature)

data InlinedEndorsementContents

Constructors

InlinedEndorsementContents

Fields

_inlinedEndorsementContents_level :: RawLevel

data OperationContentsDoubleEndorsementEvidence

"kind": { "type": "string", "enum": [ "double_endorsement_evidence" ] },

Constructors

OperationContentsDoubleEndorsementEvidence

Fields

_operationContentsDoubleEndorsementEvidence_metadata :: DoubleEndorsementEvidenceMetadata

_operationContentsDoubleEndorsementEvidence_op1 :: InlinedEndorsement

_operationContentsDoubleEndorsementEvidence_op2 :: InlinedEndorsement

data DoubleEndorsementEvidenceMetadata

Constructors

DoubleEndorsementEvidenceMetadata

Fields

_doubleEndorsementEvidenceMetadata_balanceUpdates :: !(Seq BalanceUpdate)

data OperationContentsDoubleBakingEvidence

"kind": { "type": "string", "enum": [ "double_baking_evidence" ] },

Constructors

OperationContentsDoubleBakingEvidence

Fields

_operationContentsDoubleBakingEvidence_metadata :: !DoubleBakingEvidenceMetadata

_operationContentsDoubleBakingEvidence_bh1 :: !BlockHeaderFull

_operationContentsDoubleBakingEvidence_bh2 :: !BlockHeaderFull

data DoubleBakingEvidenceMetadata

Constructors

DoubleBakingEvidenceMetadata

Fields

_doubleBakingEvidenceMetadata_balanceUpdates :: !(Seq BalanceUpdate)

data OperationContentsActivateAccount

"kind": { "type": "string", "enum": [ "activate_account" ] },

Constructors

OperationContentsActivateAccount

Fields

_operationContentsActivateAccount_metadata :: !ActivateMetadata

_operationContentsActivateAccount_pkh :: !Ed25519PublicKeyHash

_operationContentsActivateAccount_secret :: !(Base16ByteString ByteString)

data ActivateMetadata

Constructors

ActivateMetadata

Fields

_activateMetadata_balanceUpdates :: !(Seq BalanceUpdate)

data OperationContentsProposals

"kind": { "type": "string", "enum": [ "proposals" ] },

Constructors

OperationContentsProposals

Fields

_operationContentsProposals_metadata :: !EmptyMetadata

_operationContentsProposals_source :: !PublicKeyHash

_operationContentsProposals_period :: !RawLevel

_operationContentsProposals_proposals :: !(Seq ProtocolHash)

data Ballot

"ballot": { "type": "string", "enum": [ "nay", "yay", "pass" ] },

Constructors

Ballot_Nay

Ballot_Yay

Ballot_Pass

data OperationContentsBallot

"kind": { "type": "string", "enum": [ "ballot" ] },

Constructors

OperationContentsBallot

Fields

_operationContentsBallot_metadata :: !EmptyMetadata

_operationContentsBallot_source :: !PublicKeyHash

_operationContentsBallot_period :: !RawLevel

_operationContentsBallot_proposal :: !ProtocolHash

_operationContentsBallot_ballot :: !Ballot

data ManagerOperationMetadata a

Constructors

ManagerOperationMetadata

Fields

_managerOperationMetadata_balanceUpdates :: !(Seq BalanceUpdate)

_managerOperationMetadata_operationResult :: !(OperationResult a)

_managerOperationMetadata_internalOperationResults :: !(Maybe (Seq InternalOperationResult))

data InternalOperationResult

Constructors

InternalOperationResult_Reveal !InternalOperationContentsReveal

InternalOperationResult_Transaction !InternalOperationContentsTransaction

InternalOperationResult_Origination !InternalOperationContentsOrigination

InternalOperationResult_Delegation !InternalOperationContentsDelegation

data InternalOperationContentsReveal

Constructors

InternalOperationContentsReveal

data InternalOperationContentsTransaction

Constructors

InternalOperationContentsTransaction

Fields

_internalOperationContentsTransaction_source :: !ContractId

_internalOperationContentsTransaction_nonce :: !Word16

_internalOperationContentsTransaction_amount :: !Tez

_internalOperationContentsTransaction_destination :: !ContractId

_internalOperationContentsTransaction_parameters :: !(Maybe OpParameters)

_internalOperationContentsTransaction_result :: !OperationResultTransaction

data InternalOperationContentsOrigination

Constructors

InternalOperationContentsOrigination

data InternalOperationContentsDelegation

Constructors

InternalOperationContentsDelegation

data OperationResultStatus

Constructors

OperationResultStatus_Applied

OperationResultStatus_Failed

OperationResultStatus_Skipped

OperationResultStatus_Backtracked

data OperationResult a

only certain combinations of statuserrorscontent are valid, but ignore that for now

Constructors

OperationResult

Fields

_operationResult_status :: !OperationResultStatus

_operationResult_errors :: !(Maybe (Seq JsonRpcError))

_operationResult_content :: !(Maybe a)

data OperationContentsReveal

"kind": { "type": "string", "enum": [ "reveal" ] },

Constructors

OperationContentsReveal

Fields

_operationContentsReveal_metadata :: ManagerOperationMetadata OperationResultReveal

_operationContentsReveal_source :: !ContractId

_operationContentsReveal_fee :: !Tez

_operationContentsReveal_counter :: !Counter

_operationContentsReveal_gasLimit :: !TezosPositiveBigNum

_operationContentsReveal_storageLimit :: !TezosPositiveBigNum

_operationContentsReveal_publicKey :: !PublicKey

data OperationResultReveal

Constructors

OperationResultReveal

data OperationContentsTransaction

"kind": { "type": "string", "enum": [ "transaction" ] },

Constructors

OperationContentsTransaction

Fields

_operationContentsTransaction_metadata :: ManagerOperationMetadata OperationResultTransaction

_operationContentsTransaction_source :: !ContractId

_operationContentsTransaction_fee :: !Tez

_operationContentsTransaction_counter :: !Counter

_operationContentsTransaction_gasLimit :: !TezosPositiveBigNum

_operationContentsTransaction_storageLimit :: !TezosPositiveBigNum

_operationContentsTransaction_amount :: !Tez

_operationContentsTransaction_destination :: !ContractId

_operationContentsTransaction_parameters :: !(Maybe OpParameters)

data OperationResultTransaction

"operation.alpha.operation_result.transaction": {

Constructors

OperationResultTransaction

Fields

_operationResultTransaction_storage :: !(Maybe Expression)

_operationResultTransaction_balanceUpdates :: !(Seq BalanceUpdate)

_operationResultTransaction_originatedContracts :: !(Seq ContractId)

_operationResultTransaction_consumedGas :: !TezosBigNum

_operationResultTransaction_storageSize :: !TezosBigNum

_operationResultTransaction_paidStorageSizeDiff :: !TezosBigNum

data OperationContentsOrigination

"kind": { "type": "string", "enum": [ "origination" ] },

Constructors

OperationContentsOrigination

Fields

_operationContentsOrigination_metadata :: ManagerOperationMetadata OperationResultOrigination

_operationContentsOrigination_source :: !ContractId

_operationContentsOrigination_fee :: !Tez

_operationContentsOrigination_counter :: !Counter

_operationContentsOrigination_gasLimit :: !TezosPositiveBigNum

_operationContentsOrigination_storageLimit :: !TezosPositiveBigNum

_operationContentsOrigination_balance :: !Tez

_operationContentsOrigination_delegate :: !(Maybe PublicKeyHash)

_operationContentsOrigination_script :: !ContractScript

data OperationResultOrigination

Constructors

OperationResultOrigination

Fields

_operationResultOrigination_balanceUpdates :: !(Seq BalanceUpdate)

_operationResultOrigination_originatedContracts :: !(Seq ContractId)

_operationResultOrigination_consumedGas :: !TezosBigNum

_operationResultOrigination_storageSize :: !TezosBigNum

_operationResultOrigination_paidStorageSizeDiff :: !TezosBigNum

data OperationContentsDelegation

"kind": { "type": "string", "enum": [ "delegation" ] },

Constructors

OperationContentsDelegation

Fields

_operationContentsDelegation_metadata :: ManagerOperationMetadata OperationResultDelegation

_operationContentsDelegation_source :: !ContractId

_operationContentsDelegation_fee :: !Tez

_operationContentsDelegation_counter :: !Counter

_operationContentsDelegation_gasLimit :: !TezosPositiveBigNum

_operationContentsDelegation_storageLimit :: !TezosPositiveBigNum

_operationContentsDelegation_delegate :: !(Maybe PublicKeyHash)

data OperationResultDelegation

Constructors

OperationResultDelegation

stripEndorsement :: Operation -> Maybe (Op OpKind_Endorsement)

outlineEndorsement :: InlinedEndorsement -> Op OpKind_Endorsement

nonce_size :: Int

activation_code_size :: Int

jsonAddKeys :: [(Text, Value)] -> Value -> Value

checkBatch :: MonadFail m => [DSum OpKindTag OpContents] -> m (DSum OpsKindTag OpContentsList)

opParameters_value :: Lens' OpParameters Expression

opParameters_entrypoint :: Lens' OpParameters Entrypoint

seedNonceRevelationMetadata_balanceUpdates :: Iso' SeedNonceRevelationMetadata (Seq BalanceUpdate)

opContentsOrigination_script :: Lens' OpContentsOrigination ContractScript

opContentsOrigination_delegate :: Lens' OpContentsOrigination (Maybe PublicKeyHash)

opContentsOrigination_balance :: Lens' OpContentsOrigination Tez

opContentsTransaction_parameters :: Lens' OpContentsTransaction (Maybe OpParameters)

opContentsTransaction_destination :: Lens' OpContentsTransaction ContractId

opContentsTransaction_amount :: Lens' OpContentsTransaction Tez

opContentsManager_storageLimit :: forall op. Lens' (OpContentsManager op) TezosPositiveBigNum

opContentsManager_source :: forall op. Lens' (OpContentsManager op) PublicKeyHash

opContentsManager_operation :: forall op op. Lens (OpContentsManager op) (OpContentsManager op) op op

opContentsManager_gasLimit :: forall op. Lens' (OpContentsManager op) TezosPositiveBigNum

opContentsManager_fee :: forall op. Lens' (OpContentsManager op) Tez

opContentsManager_counter :: forall op. Lens' (OpContentsManager op) Counter

opContentsEndorsement_level :: Iso' OpContentsEndorsement RawLevel

op_signature :: forall a. Lens' (Op a) (Maybe Signature)

op_contents :: forall a a. Lens (Op a) (Op a) (OpContentsList a) (OpContentsList a)

op_branch :: forall a. Lens' (Op a) BlockHash

operationWithMetadata_signature :: Lens' OperationWithMetadata (Maybe Signature)

operationWithMetadata_contents :: Lens' OperationWithMetadata (Seq OperationContents)

operationResultTransaction_storageSize :: Lens' OperationResultTransaction TezosBigNum

operationResultTransaction_storage :: Lens' OperationResultTransaction (Maybe Expression)

operationResultTransaction_paidStorageSizeDiff :: Lens' OperationResultTransaction TezosBigNum

operationResultTransaction_originatedContracts :: Lens' OperationResultTransaction (Seq ContractId)

operationResultTransaction_consumedGas :: Lens' OperationResultTransaction TezosBigNum

operationResultTransaction_balanceUpdates :: Lens' OperationResultTransaction (Seq BalanceUpdate)

operationResultOrigination_storageSize :: Lens' OperationResultOrigination TezosBigNum

operationResultOrigination_paidStorageSizeDiff :: Lens' OperationResultOrigination TezosBigNum

operationResultOrigination_originatedContracts :: Lens' OperationResultOrigination (Seq ContractId)

operationResultOrigination_consumedGas :: Lens' OperationResultOrigination TezosBigNum

operationResultOrigination_balanceUpdates :: Lens' OperationResultOrigination (Seq BalanceUpdate)

operationResult_status :: forall a. Lens' (OperationResult a) OperationResultStatus

operationResult_errors :: forall a. Lens' (OperationResult a) (Maybe (Seq JsonRpcError))

operationResult_content :: forall a a. Lens (OperationResult a) (OperationResult a) (Maybe a) (Maybe a)

operationContentsTransaction_storageLimit :: Lens' OperationContentsTransaction TezosPositiveBigNum

operationContentsTransaction_source :: Lens' OperationContentsTransaction ContractId

operationContentsTransaction_parameters :: Lens' OperationContentsTransaction (Maybe OpParameters)

operationContentsTransaction_metadata :: Lens' OperationContentsTransaction (ManagerOperationMetadata OperationResultTransaction)

operationContentsTransaction_gasLimit :: Lens' OperationContentsTransaction TezosPositiveBigNum

operationContentsTransaction_fee :: Lens' OperationContentsTransaction Tez

operationContentsTransaction_destination :: Lens' OperationContentsTransaction ContractId

operationContentsTransaction_counter :: Lens' OperationContentsTransaction Counter

operationContentsTransaction_amount :: Lens' OperationContentsTransaction Tez

operationContentsSeedNonceRevelation_nonce :: Lens' OperationContentsSeedNonceRevelation (Base16ByteString ByteString)

operationContentsSeedNonceRevelation_metadata :: Lens' OperationContentsSeedNonceRevelation SeedNonceRevelationMetadata

operationContentsSeedNonceRevelation_level :: Lens' OperationContentsSeedNonceRevelation RawLevel

operationContentsReveal_storageLimit :: Lens' OperationContentsReveal TezosPositiveBigNum

operationContentsReveal_source :: Lens' OperationContentsReveal ContractId

operationContentsReveal_publicKey :: Lens' OperationContentsReveal PublicKey

operationContentsReveal_metadata :: Lens' OperationContentsReveal (ManagerOperationMetadata OperationResultReveal)

operationContentsReveal_gasLimit :: Lens' OperationContentsReveal TezosPositiveBigNum

operationContentsReveal_fee :: Lens' OperationContentsReveal Tez

operationContentsReveal_counter :: Lens' OperationContentsReveal Counter

operationContentsProposals_source :: Lens' OperationContentsProposals PublicKeyHash

operationContentsProposals_proposals :: Lens' OperationContentsProposals (Seq ProtocolHash)

operationContentsProposals_period :: Lens' OperationContentsProposals RawLevel

operationContentsProposals_metadata :: Lens' OperationContentsProposals EmptyMetadata

operationContentsOrigination_storageLimit :: Lens' OperationContentsOrigination TezosPositiveBigNum

operationContentsOrigination_source :: Lens' OperationContentsOrigination ContractId

operationContentsOrigination_script :: Lens' OperationContentsOrigination ContractScript

operationContentsOrigination_metadata :: Lens' OperationContentsOrigination (ManagerOperationMetadata OperationResultOrigination)

operationContentsOrigination_gasLimit :: Lens' OperationContentsOrigination TezosPositiveBigNum

operationContentsOrigination_fee :: Lens' OperationContentsOrigination Tez

operationContentsOrigination_delegate :: Lens' OperationContentsOrigination (Maybe PublicKeyHash)

operationContentsOrigination_counter :: Lens' OperationContentsOrigination Counter

operationContentsOrigination_balance :: Lens' OperationContentsOrigination Tez

operationContentsEndorsement_metadata :: Lens' OperationContentsEndorsement EndorsementMetadata

operationContentsEndorsement_level :: Lens' OperationContentsEndorsement RawLevel

operationContentsDoubleEndorsementEvidence_op2 :: Lens' OperationContentsDoubleEndorsementEvidence InlinedEndorsement

operationContentsDoubleEndorsementEvidence_op1 :: Lens' OperationContentsDoubleEndorsementEvidence InlinedEndorsement

operationContentsDoubleEndorsementEvidence_metadata :: Lens' OperationContentsDoubleEndorsementEvidence DoubleEndorsementEvidenceMetadata

operationContentsDoubleBakingEvidence_metadata :: Lens' OperationContentsDoubleBakingEvidence DoubleBakingEvidenceMetadata

operationContentsDoubleBakingEvidence_bh2 :: Lens' OperationContentsDoubleBakingEvidence BlockHeaderFull

operationContentsDoubleBakingEvidence_bh1 :: Lens' OperationContentsDoubleBakingEvidence BlockHeaderFull

operationContentsDelegation_storageLimit :: Lens' OperationContentsDelegation TezosPositiveBigNum

operationContentsDelegation_source :: Lens' OperationContentsDelegation ContractId

operationContentsDelegation_metadata :: Lens' OperationContentsDelegation (ManagerOperationMetadata OperationResultDelegation)

operationContentsDelegation_gasLimit :: Lens' OperationContentsDelegation TezosPositiveBigNum

operationContentsDelegation_fee :: Lens' OperationContentsDelegation Tez

operationContentsDelegation_delegate :: Lens' OperationContentsDelegation (Maybe PublicKeyHash)

operationContentsDelegation_counter :: Lens' OperationContentsDelegation Counter

operationContentsBallot_source :: Lens' OperationContentsBallot PublicKeyHash

operationContentsBallot_proposal :: Lens' OperationContentsBallot ProtocolHash

operationContentsBallot_period :: Lens' OperationContentsBallot RawLevel

operationContentsBallot_metadata :: Lens' OperationContentsBallot EmptyMetadata

operationContentsBallot_ballot :: Lens' OperationContentsBallot Ballot

operationContentsActivateAccount_secret :: Lens' OperationContentsActivateAccount (Base16ByteString ByteString)

operationContentsActivateAccount_pkh :: Lens' OperationContentsActivateAccount Ed25519PublicKeyHash

operationContentsActivateAccount_metadata :: Lens' OperationContentsActivateAccount ActivateMetadata

managerOperationMetadata_operationResult :: forall a a. Lens (ManagerOperationMetadata a) (ManagerOperationMetadata a) (OperationResult a) (OperationResult a)

managerOperationMetadata_internalOperationResults :: forall a. Lens' (ManagerOperationMetadata a) (Maybe (Seq InternalOperationResult))

managerOperationMetadata_balanceUpdates :: forall a. Lens' (ManagerOperationMetadata a) (Seq BalanceUpdate)

internalOperationContentsTransaction_source :: Lens' InternalOperationContentsTransaction ContractId

internalOperationContentsTransaction_result :: Lens' InternalOperationContentsTransaction OperationResultTransaction

internalOperationContentsTransaction_parameters :: Lens' InternalOperationContentsTransaction (Maybe OpParameters)

internalOperationContentsTransaction_nonce :: Lens' InternalOperationContentsTransaction Word16

internalOperationContentsTransaction_destination :: Lens' InternalOperationContentsTransaction ContractId

internalOperationContentsTransaction_amount :: Lens' InternalOperationContentsTransaction Tez

inlinedEndorsementContents_level :: Iso' InlinedEndorsementContents RawLevel

inlinedEndorsement_signature :: Lens' InlinedEndorsement (Maybe Signature)

inlinedEndorsement_operations :: Lens' InlinedEndorsement InlinedEndorsementContents

inlinedEndorsement_branch :: Lens' InlinedEndorsement BlockHash

endorsementMetadata_slots :: Lens' EndorsementMetadata (Seq Word8)

endorsementMetadata_delegate :: Lens' EndorsementMetadata PublicKeyHash

endorsementMetadata_balanceUpdates :: Lens' EndorsementMetadata (Seq BalanceUpdate)

doubleEndorsementEvidenceMetadata_balanceUpdates :: Iso' DoubleEndorsementEvidenceMetadata (Seq BalanceUpdate)

doubleBakingEvidenceMetadata_balanceUpdates :: Iso' DoubleBakingEvidenceMetadata (Seq BalanceUpdate)

activateMetadata_balanceUpdates :: Iso' ActivateMetadata (Seq BalanceUpdate)

noContextOperation_signature :: Lens' NoContextOperation (Maybe Signature)

noContextOperation_contents :: Lens' NoContextOperation (Seq OperationContents)

operation_signature :: Lens' Operation (Maybe Signature)

operation_protocol :: Lens' Operation ProtocolHash

operation_hash :: Lens' Operation OperationHash

operation_contents :: Lens' Operation (Seq OperationContents)

operation_chainId :: Lens' Operation ChainId

operation_branch :: Lens' Operation BlockHash

_OperationResultStatus_Backtracked :: Prism' OperationResultStatus ()

_OperationResultStatus_Skipped :: Prism' OperationResultStatus ()

_OperationResultStatus_Failed :: Prism' OperationResultStatus ()

_OperationResultStatus_Applied :: Prism' OperationResultStatus ()

_OpContentsList_Cons :: forall a k kk ks. a ~ OpKind_Manager ((:) k ((:) kk ks)) => Review (OpContentsList a) (OpContents (OpKind_Manager ((:) k ('[] :: [OpKindManager]))), OpContentsList (OpKind_Manager ((:) kk ks)))

_OpContentsList_Single :: forall a. Prism' (OpContentsList a) (OpContents a)

_InternalOperationResult_Delegation :: Prism' InternalOperationResult InternalOperationContentsDelegation

_InternalOperationResult_Origination :: Prism' InternalOperationResult InternalOperationContentsOrigination

_InternalOperationResult_Transaction :: Prism' InternalOperationResult InternalOperationContentsTransaction

_InternalOperationResult_Reveal :: Prism' InternalOperationResult InternalOperationContentsReveal

_EntrypointOther :: Prism' Entrypoint EntrypointName

_EntrypointRemoveDelegate :: Prism' Entrypoint ()

_EntrypointSetDelegate :: Prism' Entrypoint ()

_EntrypointDo :: Prism' Entrypoint ()

_EntrypointRoot :: Prism' Entrypoint ()

_EntrypointDefault :: Prism' Entrypoint ()

_OperationContents_Delegation :: Prism' OperationContents OperationContentsDelegation

_OperationContents_Origination :: Prism' OperationContents OperationContentsOrigination

_OperationContents_Transaction :: Prism' OperationContents OperationContentsTransaction

_OperationContents_Reveal :: Prism' OperationContents OperationContentsReveal

_OperationContents_Ballot :: Prism' OperationContents OperationContentsBallot

_OperationContents_Proposals :: Prism' OperationContents OperationContentsProposals

_OperationContents_ActivateAccount :: Prism' OperationContents OperationContentsActivateAccount

_OperationContents_DoubleBakingEvidence :: Prism' OperationContents OperationContentsDoubleBakingEvidence

_OperationContents_DoubleEndorsementEvidence :: Prism' OperationContents OperationContentsDoubleEndorsementEvidence

_OperationContents_SeedNonceRevelation :: Prism' OperationContents OperationContentsSeedNonceRevelation

_OperationContents_Endorsement :: Prism' OperationContents OperationContentsEndorsement

mkOp :: BlockHash -> DSum OpsKindTag OpContentsList -> Maybe Signature -> DSum OpsKindTag Op

mkSignedOp :: BlockHash -> DSum OpsKindTag OpContentsList -> Signature -> DSum OpsKindTag Op

mkUnsignedOp :: BlockHash -> DSum OpsKindTag OpContentsList -> DSum OpsKindTag Op

batchOperations :: NonEmpty (DSum BatchableOp OpContents) -> DSum OpsKindTag OpContentsList

consOpTags :: OpKindManagerTag k -> OpsKindTag (OpKind_Manager ks) -> OpsKindTag (OpKind_Manager (k ': ks))

consOpContents :: OpContents (OpKind_Manager '[k]) -> OpContentsList (OpKind_Manager ks) -> OpContentsList (OpKind_Manager (k ': ks))

batchUnfoldrM :: forall m a. Monad m => (a -> m (DSum BatchableOp OpContents, Maybe a)) -> a -> m (DSum OpsKindTag OpContentsList)