Tezos.V005.ProtocolConstants

Documentation

calcTimeBetweenBlocks :: ProtoInfo -> NominalDiffTime

data ProtoInfo

Constructors

ProtoInfo

Fields

_protoInfo_proofOfWorkNonceSize :: !Word8

_protoInfo_nonceLength :: !Word8

_protoInfo_maxRevelationsPerBlock :: !Word8

_protoInfo_maxOperationDataLength :: !Int

_protoInfo_preservedCycles :: !Cycle

_protoInfo_blocksPerCycle :: !RawLevel

_protoInfo_blocksPerCommitment :: !RawLevel

_protoInfo_blocksPerRollSnapshot :: !RawLevel

_protoInfo_blocksPerVotingPeriod :: !RawLevel

_protoInfo_timeBetweenBlocks :: !PeriodSequence

_protoInfo_endorsersPerBlock :: !Word16

_protoInfo_hardGasLimitPerOperation :: !TezosBigNum

_protoInfo_hardGasLimitPerBlock :: !TezosBigNum

_protoInfo_proofOfWorkThreshold :: !TezosInt64

_protoInfo_tokensPerRoll :: !Tez

_protoInfo_michelsonMaximumTypeSize :: !Word16

_protoInfo_seedNonceRevelationTip :: !Tez

_protoInfo_originationBurn :: !Tez

_protoInfo_originationSize :: !Int

_protoInfo_blockSecurityDeposit :: !Tez

_protoInfo_endorsementSecurityDeposit :: !Tez

_protoInfo_blockReward :: !Tez

_protoInfo_endorsementReward :: !Tez

_protoInfo_costPerByte :: !Tez

_protoInfo_hardStorageLimitPerOperation :: !TezosBigNum

_protoInfo_delayPerMissingEndorsement :: !(Maybe TezosInt64)

_protoInfo_initialEndorsers :: !(Maybe Word16)

_protoInfo_quorumMin :: !(Maybe Int)

_protoInfo_quorumMax :: !(Maybe Int)

_protoInfo_minProposalQuorum :: !(Maybe Int)

protoInfo_tokensPerRoll :: Lens' ProtoInfo Tez

protoInfo_timeBetweenBlocks :: Lens' ProtoInfo PeriodSequence

protoInfo_seedNonceRevelationTip :: Lens' ProtoInfo Tez

protoInfo_quorumMin :: Lens' ProtoInfo (Maybe Int)

protoInfo_quorumMax :: Lens' ProtoInfo (Maybe Int)

protoInfo_proofOfWorkThreshold :: Lens' ProtoInfo TezosInt64

protoInfo_proofOfWorkNonceSize :: Lens' ProtoInfo Word8

protoInfo_preservedCycles :: Lens' ProtoInfo Cycle

protoInfo_originationSize :: Lens' ProtoInfo Int

protoInfo_originationBurn :: Lens' ProtoInfo Tez

protoInfo_nonceLength :: Lens' ProtoInfo Word8

protoInfo_minProposalQuorum :: Lens' ProtoInfo (Maybe Int)

protoInfo_michelsonMaximumTypeSize :: Lens' ProtoInfo Word16

protoInfo_maxRevelationsPerBlock :: Lens' ProtoInfo Word8

protoInfo_maxOperationDataLength :: Lens' ProtoInfo Int

protoInfo_initialEndorsers :: Lens' ProtoInfo (Maybe Word16)

protoInfo_hardStorageLimitPerOperation :: Lens' ProtoInfo TezosBigNum

protoInfo_hardGasLimitPerOperation :: Lens' ProtoInfo TezosBigNum

protoInfo_hardGasLimitPerBlock :: Lens' ProtoInfo TezosBigNum

protoInfo_endorsersPerBlock :: Lens' ProtoInfo Word16

protoInfo_endorsementSecurityDeposit :: Lens' ProtoInfo Tez

protoInfo_endorsementReward :: Lens' ProtoInfo Tez

protoInfo_delayPerMissingEndorsement :: Lens' ProtoInfo (Maybe TezosInt64)

protoInfo_costPerByte :: Lens' ProtoInfo Tez

protoInfo_blocksPerVotingPeriod :: Lens' ProtoInfo RawLevel

protoInfo_blocksPerRollSnapshot :: Lens' ProtoInfo RawLevel

protoInfo_blocksPerCycle :: Lens' ProtoInfo RawLevel

protoInfo_blocksPerCommitment :: Lens' ProtoInfo RawLevel

protoInfo_blockSecurityDeposit :: Lens' ProtoInfo Tez

protoInfo_blockReward :: Lens' ProtoInfo Tez

unsafeAssumptionLevelToCycle :: ProtoInfo -> RawLevel -> Cycle

Convert a level to the cycle that contains it.

We subtract 1 because the first cycle begins after the genesis block. Yet while that block is not part of the cycle, it is still given a level, level 0.

unsafeAssumptionFirstLevelInCycle :: ProtoInfo -> Cycle -> RawLevel

Convert a cycle to the level of the first block in that cycle.

We add 1 because we do not consider the genesis block as part of the first cycle, as that would make the first cycle alone 1 block larger than all the others.

unsafeAssumptionRightsContextLevel :: ProtoInfo -> RawLevel -> RawLevel

We want the first block in the cycle that sits PRESERVED_CYCLES before the requested level, that is on the correct branch.

predictFutureTimestamp :: BlockLike blk => ProtoInfo -> RawLevel -> blk -> UTCTime